+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.0.2
 
 * Added WSM policy attachment for ESS jobs
+* Added support for 'every minute' schedules
 
 # 1.0.1